+++ imap_processing/tests/ultra/unit/conftest.py
@@ -1,7 +1,5 @@
 """Pytest plugin module for test data paths."""
 
-from unittest import mock
-
 import numpy as np
 import pytest
 import xarray as xr
@@ -19,7 +17,6 @@ from imap_processing.ultra.l0.ultra_utils import (
     ULTRA_TOF,
 )
 from imap_processing.ultra.l1a.ultra_l1a import ultra_l1a
-from imap_processing.ultra.l1b.ultra_l1b import ultra_l1b
 from imap_processing.utils import packet_file_to_datasets
 
 
@@ -209,58 +206,3 @@ def faux_aux_dataset():
     )
 
     return test_aux_dataset
-
-
-@pytest.fixture
-@mock.patch("imap_processing.ultra.l1b.de.get_annotated_particle_velocity")
-def l1b_de_dataset(
-    mock_get_annotated_particle_velocity,
-    de_dataset,
-    use_fake_spin_data_for_time,
-):
-    """L1B test data"""
-
-    data_dict = {}
-    data_dict[de_dataset.attrs["Logical_source"]] = de_dataset
-    # Create a spin table that cover spin 0-141
-    use_fake_spin_data_for_time(0, 141 * 15)
-
-    # Mock get_annotated_particle_velocity to avoid needing kernels
-    def side_effect_func(event_times, position, ultra_frame, dps_frame, sc_frame):
-        """
-        Mock behavior of get_annotated_particle_velocity.
-
-        Returns NaN-filled arrays matching the expected output shape.
-        """
-        num_events = event_times.size
-        return (
-            np.full((num_events, 3), np.nan),  # sc_velocity
-            np.full((num_events, 3), np.nan),  # sc_dps_velocity
-            np.full((num_events, 3), np.nan),  # helio_velocity
-        )
-
-    mock_get_annotated_particle_velocity.side_effect = side_effect_func
-
-    output_datasets = ultra_l1b(data_dict)
-
-    return output_datasets
-
-
-@pytest.fixture
-def l1b_extendedspin_dataset(
-    l1b_de_dataset,
-    rates_dataset,
-    faux_aux_dataset,
-):
-    """L1B de test data"""
-    data_dict = {}
-    data_dict["imap_ultra_l1b_45sensor-de"] = l1b_de_dataset[0]
-    data_dict["imap_ultra_l1a_45sensor-aux"] = faux_aux_dataset
-    # TODO: this is a placeholder for the hk dataset.
-    data_dict["imap_ultra_l1a_45sensor-hk"] = faux_aux_dataset
-    data_dict["imap_ultra_l1a_45sensor-rates"] = rates_dataset
-    data_dict["imap_ultra_l1a_45sensor-params"] = l1b_de_dataset[0]
-
-    output_datasets = ultra_l1b(data_dict)
-
-    return output_datasets
